FrontEnd/requirements.txt
os
opencv-python
Pillow
ultralytics
streamlit